+++ NEWS.md
@@ -9,6 +9,9 @@ emmeans 1.5.1.xxxx
   * `emmeans()` with contrasts specified ignores `adjust` and passes to 
     `contrast()` instead. Associated documentation improved (I hope)
   * Bug-fix for missing cases in `plot(..., comparisons = TRUE)` (#228)
+  * Robustified `plot.emmGrid()` so that comparison arrows work correctly
+    with back-transformations. (Previously we used `regrid()` in that case,
+    causing different CIs and PIs depending on `comparisons`) (#230)
   
 
 emmeans 1.5.1